requirements.txt
django==2.1.9
pillow==5.4.1
django-tinymce==2.8.0
bootstrap-admin==0.4.3
